requirements.txt
certifi
numpy
opencv-python
Pillow
python-dotenv
requests
scikit-learn
scipy
urllib3